requirements.txt
aiologger
aiohttp
pydantic
python-socketio
pyyaml
requests